extract NodeMenu
